[flask] refactor test suite
